# lib/api-zod/src/generated/types/businessBreakdownResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BusinessBreakdownResponse {
  ticker: string;
  company_name: string;
  /** Plain English explanation from Claude covering what the company does, how it makes money, and a key risk */
  explanation: string;
}